Moved plotting into a separate script, removed unsupported qc scripts
